+++ packages/ballerina-visualizer/src/views/BI/Connection/ConnectorBrowser/connectorActions.ts
@@ -18,6 +18,9 @@
 
 // Enumerating a connector's actions before any connection exists in the source. No LS
 // endpoint does this, so we read the Central docs API via library-browser/getLibraryData.
+//
+// Connector-generic: shared by the agent tool wizard and, in future, the flow diagram. Anything
+// specific to the agent tool form lives in AIChatAgent/toolForm.ts.
 
 import { AvailableNode, Category, CodeData, Item, NodeKind } from "@wso2/ballerina-core";
 import { BallerinaRpcClient } from "@wso2/ballerina-rpc-client";
@@ -64,7 +67,7 @@ export function firstSentence(value: string): string {
 }
 
 /** Mirrors `ParamUtils.REST_RESOURCE_PATH` in the LS. */
-const REST_RESOURCE_PATH = "/path/to/subdirectory";
+export const REST_RESOURCE_PATH = "/path/to/subdirectory";
 
 /** Shown instead of the LS placeholder, which reads like a real endpoint. */
 const REST_PATH_DISPLAY = "/[path...]";
@@ -277,153 +280,4 @@ export function buildConnectionSelectField(
             data: { connection: connectorCodeData },
         },
     };
-}
-
-/** Collapsible sections of the tool form. */
-export const TOOL_INPUT_GROUP = "toolInputs";
-export const RESULT_TYPE_GROUP = "resultType";
-export const OAUTH_GROUP = "oauthConfig";
-
-export const INCLUDE_CONTEXT_KEY = "includeContext";
-
-/**
- * `ai:Context ctx` becomes the tool's first parameter, so this belongs with the inputs — pass no
- * group for a form with no inputs card. A real FLAG field rather than an injected checkbox, so the
- * value arrives through the form and can sit inside a card.
- */
-export function buildIncludeContextField(group?: string): Record<string, unknown> {
-    return {
-        key: INCLUDE_CONTEXT_KEY,
-        label: "Pass agent context",
-        type: "FLAG",
-        documentation: "Adds ai:Context ctx as the first parameter so this tool can access the "
-            + "invoking agent's context.",
-        optional: true,
-        editable: true,
-        enabled: true,
-        hidden: false,
-        advanced: false,
-        value: false,
-        types: [{ fieldType: "FLAG", selected: true }],
-        ...(group ? { group } : {}),
-    };
-}
-
-interface GroupableField {
-    group?: string;
-    hidden?: boolean;
-    optional?: boolean;
-    value?: unknown;
-}
-
-export interface ToolFormGroup {
-    id: string;
-    label: string;
-    defaultCollapsed: boolean;
-}
-
-/**
- * Collapsible sections for the fields present. Inputs collapse since mappings default to
- * identity — except SQL queries, which are blanked, so hiding them would break Save.
- */
-export function buildToolFormGroups(fields: GroupableField[]): ToolFormGroup[] {
-    const visibleIn = (group: string) =>
-        fields.filter((field) => field.group === group && !field.hidden);
-
-    const inputFields = visibleIn(TOOL_INPUT_GROUP);
-    const hasUnfilledRequiredInput = inputFields.some(
-        (field) => field.optional === false && (field.value === undefined || field.value === "")
-    );
-
-    // Ordered by how often each is opened, most first.
-    const groups: ToolFormGroup[] = [];
-    if (inputFields.length > 0) {
-        groups.push({
-            id: TOOL_INPUT_GROUP,
-            label: "Inputs and Mapping",
-            defaultCollapsed: !hasUnfilledRequiredInput,
-        });
-    }
-    if (visibleIn(OAUTH_GROUP).length > 0) {
-        groups.push({ id: OAUTH_GROUP, label: "OAuth Client Configuration", defaultCollapsed: true });
-    }
-    if (visibleIn(RESULT_TYPE_GROUP).length > 0) {
-        groups.push({ id: RESULT_TYPE_GROUP, label: "Result Type", defaultCollapsed: true });
-    }
-    return groups;
-}
-
-/**
- * A resource action's symbol is only its accessor, so seed from the last named segment too.
- *
- *   `post` + `/users/[userId]/labels`      -> `postLabels`
- *   `get`  + `/users/[userId]/labels/[id]` -> `getLabels`
- */
-export function resourceToolNameSeed(accessor: string, resourcePath: string): string {
-    const path = (resourcePath || "").trim();
-    // The rest-path placeholder is not a real endpoint.
-    if (!path || path === "/" || path === REST_RESOURCE_PATH) {
-        return accessor || "";
-    }
-
-    // Skip parameters and segments `suggestToolName` would strip to nothing.
-    const named = path
-        .split("/")
-        .filter((segment) => segment && !segment.startsWith("[") && /[a-zA-Z0-9]/.test(segment));
-
-    const last = named.length > 0 ? named[named.length - 1] : "";
-    if (!last) {
-        return accessor || "";
-    }
-    return `${accessor}${last.charAt(0).toUpperCase()}${last.slice(1)}`;
-}
-
-const BALLERINA_RESERVED_TOOL_NAMES = new Set(["function", "type", "class", "service", "resource", "remote", "client"]);
-
-/** `append` -> `appendTool`, with a numeric suffix when taken. */
-export function suggestToolName(symbol: string, taken: Iterable<string>): string {
-    const existing = new Set(taken);
-    // Camel-case across separators: `get-range` -> `getRange`, not `getrange`.
-    const words = (symbol || "").split(/[^a-zA-Z0-9]+/).filter(Boolean);
-    const cleaned = words
-        .map((word, index) => (index === 0 ? word : word.charAt(0).toUpperCase() + word.slice(1)))
-        .join("");
-    if (!cleaned) {
-        return "newTool";
-    }
-    let base = `${cleaned.charAt(0).toLowerCase()}${cleaned.slice(1)}`;
-    if (!/^[a-zA-Z_]/.test(base)) {
-        base = `tool${base}`;
-    }
-    if (!base.toLowerCase().endsWith("tool")) {
-        base = `${base}Tool`;
-    }
-    if (BALLERINA_RESERVED_TOOL_NAMES.has(base)) {
-        base = `${base}Tool`;
-    }
-    if (!existing.has(base)) {
-        return base;
-    }
-    let suffix = 2;
-    while (existing.has(`${base}${suffix}`)) {
-        suffix++;
-    }
-    return `${base}${suffix}`;
-}
-
-/** Names already used by the agent's tools, so a suggestion never collides. */
-export function getExistingToolNames(agentNode: { properties?: Record<string, any> } | undefined): string[] {
-    const raw = agentNode?.properties?.tools?.value;
-    if (Array.isArray(raw)) {
-        return raw.map((entry) => String(entry).trim()).filter(Boolean);
-    }
-    if (typeof raw !== "string") {
-        return [];
-    }
-    return raw
-        .replace(/^\s*\[/, "")
-        .replace(/\]\s*$/, "")
-        .split(",")
-        .map((entry) => entry.trim())
-        .filter(Boolean);
 }
