+++ packages/ballerina-visualizer/src/views/BI/Connection/ConnectorBrowser/index.ts
@@ -1,0 +1,37 @@
+/**
+ * Copyright (c) 2026, WSO2 LLC. (https://www.wso2.com) All Rights Reserved.
+ *
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+// Browsing a connector's actions before any connection exists. Used by the agent tool wizard;
+// intended for the flow diagram's connection flow too.
+
+export { ConnectorBrowser, WizardStep } from "./ConnectorBrowser";
+export type { ActionSelection } from "./ConnectorBrowser";
+export { ConnectorActionList } from "./ConnectorActionList";
+export { ConnectorList } from "./ConnectorList";
+export {
+    actionDisplayLabel,
+    buildConnectionSelectField,
+    displayResourcePath,
+    fetchConnectorActions,
+    firstSentence,
+    formatActionLabel,
+    formatResourceSignature,
+    normalizeConnectorSearchCategories,
+    REST_RESOURCE_PATH,
+    toResourcePathTemplate,
+} from "./connectorActions";
